% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{execute_mutation}
\alias{execute_mutation}
\title{Execute a GraphQL mutation}
\usage{
execute_mutation(
  client,
  mutation,
  mutation_name = "mutation",
  variables = NULL
)
}
\arguments{
\item{client}{GraphQL client object from create_gql_client}

\item{mutation}{GraphQL mutation string}

\item{mutation_name}{Name for the mutation (default: "mutation")}

\item{variables}{Optional variables list}
}
\value{
Mutation result
}
\description{
Execute a GraphQL mutation
}
