% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{execute_query}
\alias{execute_query}
\title{GraphQL Operations}
\usage{
execute_query(client, query, query_name = "query", variables = NULL)
}
\arguments{
\item{client}{GraphQL client object from create_gql_client}

\item{query}{GraphQL query string}

\item{query_name}{Name for the query (default: "query")}

\item{variables}{Optional variables list}
}
\value{
Query result
}
\description{
Functions for executing GraphQL queries and mutations.
Execute a GraphQL query
}
