% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth.R
\name{get_oauth_token}
\alias{get_oauth_token}
\title{OAuth 2 Authentication Functions}
\usage{
get_oauth_token(client_id, client_secret, token_url, token_file)
}
\arguments{
\item{client_id}{OAuth 2 client ID}

\item{client_secret}{OAuth 2 client secret}

\item{token_url}{OAuth 2 token URL}

\item{token_file}{Path to file to store token}
}
\value{
OAuth token object
}
\description{
Functions for obtaining OAuth 2 tokens using different grant types.
}
\details{
For authorization code flows, the \code{httpuv} package is recommended for better
OAuth callback handling. Install it with: \code{install.packages("httpuv")}
Get OAuth 2 token (client credentials flow)
}
