% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth.R
\name{get_oauth_token_authorization_code}
\alias{get_oauth_token_authorization_code}
\title{Get OAuth 2 token using Authorization Code flow}
\usage{
get_oauth_token_authorization_code(
  client_id,
  client_secret,
  authorize_url = "https://app.playerdata.co.uk/oauth/authorize",
  token_url = "https://app.playerdata.co.uk/oauth/token",
  token_file = ".token",
  port = 8080,
  redirect_uri = NULL
)
}
\arguments{
\item{client_id}{OAuth 2 client ID}

\item{client_secret}{OAuth 2 client secret}

\item{authorize_url}{OAuth 2 authorization URL (default: https://app.playerdata.co.uk/oauth/authorize)}

\item{token_url}{OAuth 2 token URL (default: https://app.playerdata.co.uk/oauth/token)}

\item{token_file}{Path to file to store token (default: .token)}

\item{port}{Port for local callback server (default: 8080)}

\item{redirect_uri}{Redirect URI (default: http://localhost:port/callback)}
}
\value{
OAuth token object
}
\description{
Get OAuth 2 token using Authorization Code flow
}
