% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parseQueryString}
\alias{parseQueryString}
\title{Utility Functions}
\usage{
parseQueryString(query_string)
}
\arguments{
\item{query_string}{Query string (e.g., "code=abc&state=xyz")}
}
\value{
Named list of query parameters
}
\description{
Parse query string from URL
}
