% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{refresh_client_token}
\alias{refresh_client_token}
\title{Refresh the OAuth token for a client}
\usage{
refresh_client_token(client)
}
\arguments{
\item{client}{GraphQL client object}
}
\description{
Refresh the OAuth token for a client
}
